Corrected message sizes in the example message size table
Thank to Loïc Ferreira for pointing out that the table number 116 did not argree with the size 115 bytes in the recent -10 test vectors.

I looked though the details in -10 and earlier version. 115 is correct for version -10. 117 was correct for earlier versions. The concatenation of G_Y and CIPHERTEXT_2 in -10 lowered the size from 117 to 115. The number 116 in -08 -09 and -10 was incorrect (the change in -08 affected message_3 but not message_2).
 
message_2                    115 bytes
   G_Y_CIPHERTEXT_2          114 bytes
      bstr header              2 bytes
      G_Y                     32 bytes
      CIPHERTEXT_2            80 bytes
         ID_CRED_R            14 bytes
         Signature_or_MAC_2   66 bytes
   C_R                         1 bytes