Updates of Figure 2 and 3
Figure 2 used a different format from the other 3 message flow figures.
Figure 2 showed a AEAD variant that is not described in the SIGMA paper or EDHOC. I think it might be more confusing than helpful for the reader.
Figure 3 should likely show message_4. Unless we think that will be confusing and people will think that message_4 is normally used.