Update draft-ietf-lake-edhoc.md
Fixed an error where the ead section said that the data was an int.
Added value type as suggested in #166 
Hard to combined the different terms here in a good way. I found that label, value, and value type seemed best. That aligns with some (but not all) COSE and CWT registries.