MTI section #73
Suggestion base the discussion during the October interim. There was no strong opinions. This PR updates the MTI for header parameters and credential types to align with the choices made for method (none is MTI).

EAD is explicitly optional to support. It is used in specific cases and requires additional mechanics for each use case.

#73