Unicode for Mathematical operators and symbols
Asterisk * for multiplication and <= >= make it looks like somebody wrote the text on a typewriter. A blast from the past in a bad way. Asterisk is also used for at least three other meanings in the document.

Changed to ∙, ≤, and ≥

My understanding is this that it is now allowed to use unicode as above (and if not the RFC will change it back against my will)